requirements.txt
Django
channels==4.2.0
constantly==23.10.4
cryptography==42.0.5
daphne==4.1.2
decouple==0.0.7
dj-database-url==2.1.0
django-cors-headers==4.6.0
django-oauth-toolkit==3.0.1
django-rest-framework-social-oauth2==1.2.0
django-social-auth==0.7.28
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-social-oauth2==3.1.0
drf-yasg==1.21.7
gunicorn==22.0.0
Markdown==3.6
oauth2==1.9.0.post1
oauth2_provider==0.0
pillow==10.3.0
channels-redis
psycopg2-binary==2.9.9
python-dotenv==1.0.1
redis==5.0.4
requests==2.31.0
six==1.16.0
social-auth-app-django==5.4.0
whitenoise==6.6.0